Improve presentation (#188)
* feat: Init Contribution and fill Readme

* docs: Add illustrations to readme, write contribution

Add more details on the various sections of the ReadMe and illustrations. Complete the Contribution file with descriptions and more guidelines. Contains changes for milestones 2.

* docs: Merge CONTRIBUTING and README

To start small, I don't think I need a CONTRIBUTING, CODE_OF_CONDUCT and CHANGELOG files. All should fit in the README for now.

* refactor: Update Readme with link to CC docs.

Also remove the `thumb up` and `thumb down` icons, as they are not completed yet.

* fix: Remove references to the Figma UI Kit.

The Figma Kit will be linked back here when it will be published on the Community.

* refactor: Update readme

* refactor: Update ReadMe

Change some illustration from PNG to SVG.

* fix: backlinks not working on github

* fix: backlink not respecting file Capitalization

* style(identity): decorate headers with icons

Replace emojis with icons (because why not)

* fix(identity): remove inline css in imgs

Github eat custom inline CSS in the style attributes to put its own. It's better to have icons on the right for now.

* fix(identity): fix a header issue in ReadMe

* fix(identity): Reduce icons size to match github headers hights

* Revert "fix(identity): Reduce icons size to match github headers hights"

This reverts commit 2ae463bdbbe5d59ea1de746372e3978f272cafc5.